+++ src/vector2.hpp
@@ -7,15 +7,15 @@
 namespace green_leaf {
   class Vector2 {
   public:
-    Vector2(float x, float y);
+    Vector2(int x, int y);
 
     std::string toString() const;
 
-    float x() const {
+    int x() const {
       return x_;
     }
 
-    float y() const {
+    int y() const {
       return y_;
     }
 
@@ -29,14 +29,14 @@ namespace green_leaf {
 
     Vector2 operator+(const Vector2& p) const;
     Vector2 operator-(const Vector2& p) const;
-    Vector2 operator*(float f) const;
+    Vector2 operator*(int f) const;
     Vector2 operator*(const Vector2& p) const;
-    Vector2 operator/(float f) const;
+    Vector2 operator/(int f) const;
     Vector2 operator/(const Vector2& p) const;
 
   private:
-    float x_;
-    float y_;
+    int x_;
+    int y_;
   };
 
   ::std::ostream& operator<<(::std::ostream& os, const Vector2& v);
@@ -45,7 +45,7 @@ namespace green_leaf {
 namespace std {
   template <> struct hash<green_leaf::Vector2> {
     size_t operator()(const green_leaf::Vector2 &v) const {
-      std::hash<float> hasher;
+      std::hash<int> hasher;
       return hasher(v.x()) ^ hasher(v.y());
     }
   };
